+++ src/app/locales/en/components/new-cluster-zone-registration.js
@@ -1,0 +1,3 @@
+export default {
+  providerRegisteredSuccessfully: 'Provider registered successfully',
+};