+++ src/lib/ember-onedata-onepanel-server/addon/utils/response-validators/oneprovider/get-provider.js
@@ -14,8 +14,7 @@ export default function (data) {
     return !!(
       data.id &&
       data.name &&
-      Array.isArray(data.urls) &&
-      data.redirectionPoint
+      typeof data.subdomainDelegation === 'boolean'
     );
   } catch (error) {
     if (error instanceof TypeError) {