+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-3606 Subdomain delegation functionality for provider
 * VFS-3661 Improvements in presenting loading state of views and errors
 * VFS-3685 Added space files popularity and space auto cleaning views
     * redesigned single space view to use tabs (sync. stats, files pop., space auto cleaning)